Add if statement to zebrafyChildren method
